move first and last
