+++ addon/array/first.js
@@ -1,7 +1,7 @@
-import { normalizeArray } from './utils';
+import { normalizeArray } from '../utils';
 
 export default function(...keys) {
   return normalizeArray(keys, {}, array => {
-    return array[array.length - 1];
+    return array[0];
   });
 }
