+++ addon/array/index.js
@@ -6,7 +6,9 @@ export { default as filterBy } from './filter-by';
 export { default as filter } from './filter';
 export { default as findBy } from './find-by';
 export { default as find } from './find';
+export { default as first } from './first';
 export { default as includes } from './includes';
 export { default as indexOf } from './index-of';
 export { default as lastIndexOf } from './last-index-of';
+export { default as last } from './last';
 export { default as length } from './length';
